+++ src/app/(frontend)/[center]/blog/page.tsx
@@ -4,9 +4,8 @@ import { PostsList } from '@/components/PostsList'
 import { FiltersTotalProvider } from '@/contexts/FiltersTotalContext'
 import type { Metadata, ResolvedMetadata } from 'next/types'
 import { createLoader, parseAsString, SearchParams } from 'nuqs/server'
-import { BlogMobileFilters } from './blog-mobile-filters'
-import { BlogTagsFilter } from './blog-tags-filter'
-import { PostsSort } from './posts-sort'
+import { BlogMobileFilters } from './BlogMobileFilters'
+import { PostsFilters } from './PostsFilters'
 
 const blogSearchParams = {
   sort: parseAsString.withDefault('-publishedAt'),
@@ -43,13 +42,8 @@ export default async function Page({ params, searchParams }: Args) {
       <div className="pt-4">
         <div className="container md:max-lg:max-w-5xl mb-16 flex flex-col-reverse md:flex-row flex-1 gap-10 md:gap-16">
           <div className="grow">
-            {/* Mobile filters drawer */}
             <div className="md:hidden mb-4">
-              <BlogMobileFilters
-                tags={tagsList}
-                initialSort={sort}
-                hasActiveFilters={hasActiveFilters}
-              />
+              <BlogMobileFilters tags={tagsList} sort={sort} hasActiveFilters={hasActiveFilters} />
             </div>
 
             <PostsList
@@ -63,8 +57,7 @@ export default async function Page({ params, searchParams }: Args) {
 
           {/* Desktop sorting and filters */}
           <div className="hidden md:flex flex-col shrink-0 justify-between md:justify-start md:w-[240px] lg:w-[300px]">
-            <PostsSort initialSort={sort} />
-            {tagsList.length > 1 && <BlogTagsFilter tags={tagsList} />}
+            <PostsFilters sort={sort} tags={tagsList} />
           </div>
         </div>
       </div>
